requirements.txt
# Core dependencies
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
pydantic>=2.5.0
python-dateutil>=2.8.2

# LLM support (optional but recommended)
openai>=1.3.0
anthropic>=0.7.0

# NLP utilities
spacy>=3.7.0
regex>=2023.10.3

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
